final final real final wiggle final real actual wiggle real final final

+++ AcoBurd/lib/BurdLib/include/timers.hpp
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <stdint.h>
+
 #include "get_set_macros.hpp"
 
 // Timer Add Configuration
@@ -14,14 +16,14 @@
 #define LORA_INTERVAL 60                      // Time between LoRa transmits
 #define GPS_INTERVAL 3600                     // Time between GPS updates
 
-GET_SET_FUNC_PROTO(long, lora_timer)
-GET_SET_FUNC_PROTO(long, display_timer)
-GET_SET_FUNC_PROTO(long, encoder_timer)
-GET_SET_FUNC_PROTO(long, release_timer)
-GET_SET_FUNC_PROTO(long, release_timer1)
-GET_SET_FUNC_PROTO(long, release_timer2)
-GET_SET_FUNC_PROTO(long, battery_timer)
-GET_SET_FUNC_PROTO(long, time_until_release)
+GET_SET_FUNC_PROTO(uint32_t, lora_timer)
+GET_SET_FUNC_PROTO(uint32_t, display_timer)
+GET_SET_FUNC_PROTO(uint32_t, encoder_timer)
+GET_SET_FUNC_PROTO(uint32_t, release_timer)
+GET_SET_FUNC_PROTO(uint32_t, release_timer1)
+GET_SET_FUNC_PROTO(uint32_t, release_timer2)
+GET_SET_FUNC_PROTO(uint32_t, battery_timer)
+GET_SET_FUNC_PROTO(uint32_t, time_until_release)
 
 void
 release_service(
